Finally fix the problem and we have sound
Turns out that the channles keep outputting the last
output they outputed so we have more samples than we thought.

+++ lib_gb/src/apu/wave_sample_producer.rs
@@ -18,7 +18,7 @@ impl Default for WaveSampleProducer{
 }
 
 impl SampleProducer for WaveSampleProducer{
-    fn produce(&mut self) ->u8 {
+    fn produce(&mut self) ->i8 {
         let mut sample = self.wave_samples[(self.sample_counter/2) as usize];
 
         if self.sample_counter % 2 != 0{
@@ -35,7 +35,7 @@ impl SampleProducer for WaveSampleProducer{
             self.sample_counter = 0;
         }
 
-        return self.shift_by_volume(sample);
+        return self.shift_by_volume(sample) as i8;
     }
 }
